# Conflicts:
#	src/main/webapp/WEB-INF/web.xml
#	src/main/webapp/view/customer-home.jsp

+++ src/main/java/model/Order.java
@@ -84,7 +84,7 @@ public class Order {
         return orderTotalAmount;
     }
 
-    public void setOrderTotalAmount(int orderTotalAmount) {
+    public void setOrderTotalAmount(double orderTotalAmount) {
         this.orderTotalAmount = orderTotalAmount;
     }
